requirements.txt
# Telegram Bot
python-telegram-bot==21.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Scheduling
apscheduler==3.10.4

# Environment variables
python-dotenv==1.0.0

# Timezone
pytz==2024.1
